Fix bernoulli sampling.

+++ noise-2-noise/args.py
@@ -48,7 +48,7 @@ def get_args(batch_size=16, ih=256, iw=256, max_iter=468750, save_interval=3125)
     parser.add_argument("--noise-dist", type=str, default="gaussian",
                         help="Noise distribution", 
                         choices=["gaussian", "poisson", "bernoulli"])
-    parser.add_argument("--noise-level", type=int, default=50,
+    parser.add_argument("--noise-level", type=float, default=50,
                         help="Noise level")
     parser.add_argument("--save-interval", type=int, default=3125,
                         help="Interval for saving models.")
